LayoutTests/editing/selection/update-selection-with-modify-dissociates-range-expected.txt

This tests calling getSelection().modify to update selection. getSelection().getRangeAt(0) should return a distinct range after each modification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range1 != range2 is true
PASS range2 != range3 is true
PASS range3 != range4 is true
PASS successfullyParsed is true

TEST COMPLETE
hello, world
